Added manufacturer "_TZ3000_b28wrpvx" (BSeed, Wall Outlet with metering)
The "_TZ3000_b28wrpvx" Wall Outlet works according the smartplug. The Wall Outlet is strictly speaking not a plug, but the metering seems to be fairly accurate and it works fine for already a month or so. As one would have a specific device plugged in (a washing machine or dryer) one can always change the icon.